=General *Merged changes from 3ceam-2 rev. 19
*added some comments here and there

*Increased limit from 32,767 to 65,535, thanks to rAthena Devs.
-Still uses same amount of memory with doubled usable size.
-Adjusted CARD0_PET from -256 to 256.
-Please run both SQL files after applying this update to avoid values becoming corrupt.

*Added support for the transformation scrolls.
-Transformation scrolls changes a player into a monster for 20 minutes when used.

*Added a huge number of SC's to the const.txt and status.h file.
-This is done to prepare for the coding of the skills for 3rd jobs and other jobs.
-The tags and ID's assigned to them are now finalized and will not be changed.

*Updated some OPT1 checks so the burning status will not lock the affected player.

*Fixed a suspected issue where the Moon Star's animation would vanish if the player dies or gets hit with a debuff like effect.

*Added full support for the Super Star item.

=Database
*Updated the transformation scrolls.

+++ src/map/clif.h
@@ -452,7 +452,7 @@ void clif_skill_teleportmessage(struct map_session_data *sd, int type);
 void clif_skill_produce_mix_list(struct map_session_data *sd, int trigger);
 void clif_cooking_list(struct map_session_data *sd, int trigger);
 
-void clif_produceeffect(struct map_session_data* sd,int flag,int nameid);
+void clif_produceeffect(struct map_session_data* sd,int flag,unsigned short nameid);
 
 void clif_skill_setunit(struct skill_unit *unit);
 void clif_skill_delunit(struct skill_unit *unit);
@@ -497,7 +497,7 @@ void clif_item_refine_list(struct map_session_data *sd);
 void clif_item_skill(struct map_session_data *sd,int skillid,int skilllv);
 
 void clif_mvp_effect(struct map_session_data *sd);
-void clif_mvp_item(struct map_session_data *sd,int nameid);
+void clif_mvp_item(struct map_session_data *sd,unsigned short nameid);
 void clif_mvp_exp(struct map_session_data *sd, unsigned int exp);
 void clif_mvp_noitem(struct map_session_data* sd);
 void clif_changed_dir(struct block_list *bl, enum send_target target);
@@ -594,7 +594,7 @@ void clif_map_property(struct map_session_data* sd, enum map_property property);
 void clif_pvpset(struct map_session_data *sd, int pvprank, int pvpnum,int type);
 void clif_map_property_mapall(int map, enum map_property property);
 void clif_refine(int fd, int fail, int index, int val);
-void clif_upgrademessage(int fd, int result, int item_id);
+void clif_upgrademessage(int fd, int result, unsigned short item_id);
 
 //petsystem
 void clif_catch_process(struct map_session_data *sd);
@@ -698,8 +698,8 @@ void clif_mercenary_message(struct map_session_data* sd, int message);
 void clif_mercenary_updatestatus(struct map_session_data *sd, int type);
 
 // RENTAL SYSTEM
-void clif_rental_time(int fd, int nameid, int seconds);
-void clif_rental_expired(int fd, int index, int nameid);
+void clif_rental_time(int fd, unsigned short nameid, int seconds);
+void clif_rental_expired(int fd, int index, unsigned short nameid);
 
 // BOOK READING
 void clif_readbook(int fd, int book_id, int page);
