+++ src/map/log.c
@@ -90,7 +90,7 @@ static char log_chattype2char(e_log_chat_type type)
 
 
 /// check if this item should be logged according the settings
-static bool should_log_item(int nameid, int amount, int refine)
+static bool should_log_item(unsigned short nameid, int amount, int refine)
 {
 	int filter = log_config.filter;
 	struct item_data* id;
@@ -158,7 +158,7 @@ void log_branch(struct map_session_data* sd)
 
 
 /// logs item transactions
-void log_pick(struct block_list* bl, e_log_pick_type type, int nameid, int amount, struct item* itm)
+void log_pick(struct block_list* bl, e_log_pick_type type, unsigned short nameid, int amount, struct item* itm)
 {
 	const char* mapname;
 	int id = 0;
@@ -195,7 +195,7 @@ void log_pick(struct block_list* bl, e_log_pick_type type, int nameid, int amoun
 	{
 		if( itm == NULL )
 		{//We log common item
-			if( SQL_ERROR == Sql_Query(logmysql_handle, "INSERT DELAYED INTO `%s` (`time`, `char_id`, `type`, `nameid`, `amount`, `map`) VALUES (NOW(), '%d', '%c', '%d', '%d', '%s')",
+			if( SQL_ERROR == Sql_Query(logmysql_handle, "INSERT DELAYED INTO `%s` (`time`, `char_id`, `type`, `nameid`, `amount`, `map`) VALUES (NOW(), '%d', '%c', '%hu', '%d', '%s')",
 				log_config.log_pick, id, log_picktype2char(type), nameid, amount, mapname) )
 			{
 				Sql_ShowDebug(logmysql_handle);
@@ -204,7 +204,7 @@ void log_pick(struct block_list* bl, e_log_pick_type type, int nameid, int amoun
 		}
 		else
 		{//We log Extended item
-			if( SQL_ERROR == Sql_Query(logmysql_handle, "INSERT DELAYED INTO `%s` (`time`, `char_id`, `type`, `nameid`, `amount`, `refine`, `card0`, `card1`, `card2`, `card3`, `map`) VALUES (NOW(), '%d', '%c', '%d', '%d', '%d', '%d', '%d', '%d', '%d', '%s')",
+			if( SQL_ERROR == Sql_Query(logmysql_handle, "INSERT DELAYED INTO `%s` (`time`, `char_id`, `type`, `nameid`, `amount`, `refine`, `card0`, `card1`, `card2`, `card3`, `map`) VALUES (NOW(), '%d', '%c', '%hu', '%d', '%d', '%hu', '%hu', '%hu', '%hu', '%s')",
 				log_config.log_pick, id, log_picktype2char(type), itm->nameid, amount, itm->refine, itm->card[0], itm->card[1], itm->card[2], itm->card[3], mapname) )
 			{
 				Sql_ShowDebug(logmysql_handle);
@@ -226,11 +226,11 @@ void log_pick(struct block_list* bl, e_log_pick_type type, int nameid, int amoun
 
 		if( itm == NULL )
 		{//We log common item
-			fprintf(logfp,"%s - %d\t%c\t%d,%d,%s\n", timestring, id, log_picktype2char(type), nameid, amount, mapname);
+			fprintf(logfp,"%s - %d\t%c\t%hu,%d,%s\n", timestring, id, log_picktype2char(type), nameid, amount, mapname);
 		}
 		else
 		{//We log Extended item
-			fprintf(logfp,"%s - %d\t%c\t%d,%d,%d,%d,%d,%d,%d,%s\n", timestring, id, log_picktype2char(type), itm->nameid, amount, itm->refine, itm->card[0], itm->card[1], itm->card[2], itm->card[3], mapname);
+			fprintf(logfp,"%s - %d\t%c\t%hu,%d,%d,%hu,%hu,%hu,%hu,%s\n", timestring, id, log_picktype2char(type), itm->nameid, amount, itm->refine, itm->card[0], itm->card[1], itm->card[2], itm->card[3], mapname);
 		}
 		fclose(logfp);
 	}
